interval to double

+++ src/main/java/com/websocket/kafka/example/websocket/model/MetricInfo.java
@@ -14,7 +14,7 @@ public class MetricInfo {
     private HashMap<String,String> descriptions;
     private String dataType;
     private int max;
-    private int interval;
+    private Double interval;
 
     public MetricInfo(){}
 }
